+++ sound-mind-studio/tests/test_filter_configuration_panel.h
@@ -20,4 +20,8 @@ private slots:
     void selectingToneCurveShowsItsOwnGroup();
     void editingTheToneCurveEditorUpdatesConfigAndEmits();
     void setFilterConfigurationSyncsTheToneCurveEditorWithoutEmitting();
+    void equalizerModeHidesTheFilterTypeComboAndShowsTheCutGroup();
+    void equalizerModeOffRestoresTheNormalPerTypeGroup();
+    void editingACutSpinBoxWritesOpacityAndForcesIntensityToTheSilenceFloor();
+    void setFilterConfigurationSyncsTheCutSpinBoxesFromOpacityWithoutEmitting();
 };